Parse markdown links and newlines in app descriptions
